Fixed a three nits
These were caught in earlier reviews.

+++ launch/launch-node.py
@@ -298,7 +298,8 @@ def main():
     dns.print_dns(client, options.name)
     # Remove the ansible inventory cache so that next run finds the new
     # server
-    os.unlink('/var/cache/ansible-inventory.json')
+    if os.path.exists('/var/cache/ansible-inventory.json'):
+        os.unlink('/var/cache/ansible-inventory.json')
 
 if __name__ == '__main__':
     main()
